+++ src/Factory/AbstractFactoryInterface.php
@@ -15,8 +15,8 @@ use Interop\Container\ContainerInterface;
  * Interface for an abstract factory.
  *
  * An abstract factory extends the factory interface, but also has an
- * additional "canCreateService" method, which is called to check if the
- * abstract factory can create an instance for the given service. You should
+ * additional "has" method, which is called to check if the abstract factory
+ * has the ability to create an instance for the given service. You should
  * limit the number of abstract factories to ensure good performance. Starting
  * from ServiceManager v3, remember that you can also attach multiple names to
  * the same factory, which reduces the need for abstract factories.
@@ -30,5 +30,5 @@ interface AbstractFactoryInterface extends FactoryInterface
      * @param  string $requestedName
      * @return bool
      */
-    public function canCreateServiceWithName(ContainerInterface $container, $requestedName);
+    public function canCreate(ContainerInterface $container, $requestedName);
 }